V0_10_5
Fix sniff bug
InitializeExitSniffOnReceive() was initializing nvEnableSniffEntry instead of NvExitSniffOnReceive which prevented the stack from asking to exit sniff when a message was received from the phone

Initialize the number of free entries to the stack depth for the task debugging information

add messages.h to the project file view in IAR

Default is to use the CC2560 patch for all radios.  This should be a temporary change.