Fix vibration cycle count.
In firmware 0.8.X, a cycle count of 1 would mean 1 vibration. With the current implementation, a cycle count of 1 results in 2 vibrations. This causes confusion for both users and developers and is not really backwards compatible.

This commit changes the behavior to the old, expected one, with a very minor change.